# WMS.Angular/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "WMS API v1",
    "version": "v1"
  },
  "paths": {
...
    "/receiving": {
      "get": {
        "tags": [
          "Receiving"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReceivingSummaryDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Receiving"
        ],
        "summary": "Create a new receiving receipt",
        "description": "Creates a receiving receipt linked to an Incoming shipment, auto-generates sequential lot numbers per product, marks specific line items as received, and updates Incoming status to PARTIAL or RECEIVED.",
        "operationId": "CreateReceiving",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReceivingDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReceivingDetailsDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/receiving/shippers": {
      "get": {
        "tags": [
          "Receiving"
        ],
        "summary": "Get all unique shippers from receiving receipts",
        "description": "Retrieves a distinct list of non-empty shipper names across all receiving records, sorted alphabetically.",
        "operationId": "GetAllShippersFromReceiving",
        "parameters": [
          {
            "name": "warehouseId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/receiving/warehouse/{WarehouseId}": {
      "get": {
        "tags": [
          "Receiving"
        ],
        "parameters": [
          {
            "name": "WarehouseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReceivingSummaryDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/receiving/{id}": {
      "get": {
        "tags": [
          "Receiving"
        ],
        "summary": "Get receiving transaction details",
        "description": "Retrieves full receiving receipt details by ID, including logistics header info, linked warehouse, assigned pallets, and line-item products with their server-generated lot numbers.",
        "operationId": "GetReceiving",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReceivingDetailsDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Receiving"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReceivingDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Receiving"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/warehouse/{id}": {
      "get": {
        "tags": [
          "WMS.Api"
        ],
        "operationId": "GetWarehouse",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "WMS.Api"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWarehouseDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "WMS.Api"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateReceivingDto": {
        "type": "object",
        "properties": {
          "warehouseId": {
            "type": "integer",
            "format": "int32"
          },
          "incomingId": {
            "type": "integer",
            "format": "int32"
          },
          "series": {
            "type": "string",
            "nullable": true
          },
          "transportCompany": {
            "type": "string",
            "nullable": true
          },
          "shipper": {
            "type": "string",
            "nullable": true
          },
          "consignee": {
            "type": "string",
            "nullable": true
          },
          "dateReceived": {
            "type": "string",
            "format": "date-time"
          },
          "reference": {
            "type": "string",
            "nullable": true
          },
          "products": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReceivedProductDetailsDto"
            },
            "nullable": true
          },
          "plateNumber": {
            "type": "string",
            "nullable": true
          },
          "driverName": {
            "type": "string",
            "nullable": true
          },
          "dateTime": {
            "type": "string",
            "format": "date-time"
          },
          "clientRepresentative": {
            "type": "string",
            "nullable": true
          },
          "checkerName": {
            "type": "string",
            "nullable": true
          },
          "timeStart": {
            "type": "string",
            "format": "date-time"
          },
          "timeEnd": {
            "type": "string",
            "format": "date-time"
          },
          "dateAdded": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "nullable": true
          }
        },
        "additionalProperties": {}
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": {}
      },
...
      "ReceivedProductSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "productId": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "lotNumber": {
            "type": "string",
            "nullable": true
          },
          "typeOfPackage": {
            "type": "string",
            "nullable": true
          },
          "measurement": {
            "type": "string",
            "nullable": true
          },
          "weight": {
            "type": "number",
            "format": "double"
          },
          "expectedProductName": {
            "type": "string",
            "nullable": true
          },
          "expectedQuantity": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "expectedCBM": {
            "type": "string",
            "nullable": true
          },
          "expectedTotalWeight": {
            "type": "string",
            "nullable": true
          },
          "expectedExpirationDate": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "cbm": {
            "type": "string",
            "nullable": true
          },
          "totalWeight": {
            "type": "string",
            "nullable": true
          },
          "expirationDate": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "remarks": {
            "type": "string",
            "nullable": true
          },
          "containerName": {
            "type": "string",
            "nullable": true
          },
          "palletId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "variance": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "hasDiscrepancy": {
            "type": "boolean",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "ReceivingDetailsDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "warehouseId": {
            "type": "integer",
            "format": "int32"
          },
          "incomingId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "series": {
            "type": "string",
            "nullable": true
          },
          "transportCompany": {
            "type": "string",
            "nullable": true
          },
          "shipper": {
            "type": "string",
            "nullable": true
          },
          "consignee": {
            "type": "string",
            "nullable": true
          },
          "dateReceived": {
            "type": "string",
            "format": "date-time"
          },
          "reference": {
            "type": "string",
            "nullable": true
          },
          "products": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReceivedProductDetailsDto"
            },
            "nullable": true
          },
          "plateNumber": {
            "type": "string",
            "nullable": true
          },
          "driverName": {
            "type": "string",
            "nullable": true
          },
          "dateTime": {
            "type": "string",
            "format": "date-time"
          },
          "clientRepresentative": {
            "type": "string",
            "nullable": true
          },
          "checkerName": {
            "type": "string",
            "nullable": true
          },
          "timeStart": {
            "type": "string",
            "format": "date-time"
          },
          "timeEnd": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "ReceivingSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "warehouseId": {
            "type": "integer",
            "format": "int32"
          },
          "warehouse": {
            "type": "string",
            "nullable": true
          },
          "series": {
            "type": "string",
            "nullable": true
          },
          "transportCompany": {
            "type": "string",
            "nullable": true
          },
          "shipper": {
            "type": "string",
            "nullable": true
          },
          "consignee": {
            "type": "string",
            "nullable": true
          },
          "dateReceived": {
            "type": "string",
            "format": "date-time"
          },
          "reference": {
            "type": "string",
            "nullable": true
          },
          "products": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReceivedProductSummaryDto"
            },
            "nullable": true
          },
          "plateNumber": {
            "type": "string",
            "nullable": true
          },
          "driverName": {
            "type": "string",
            "nullable": true
          },
          "dateTime": {
            "type": "string",
            "format": "date-time"
          },
          "clientRepresentative": {
            "type": "string",
            "nullable": true
          },
          "checkerName": {
            "type": "string",
            "nullable": true
          },
          "timeStart": {
            "type": "string",
            "format": "date-time"
          },
          "timeEnd": {
            "type": "string",
            "format": "date-time"
          },
          "dateAdded": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "UpdateWarehouseDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "floorPlanImage": {
            "type": "string",
            "format": "byte",
            "nullable": true
          },
          "dateAdded": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "CookieAuth": {
        "type": "apiKey",
        "description": "ASP.NET Core Identity Application Cookie",
        "name": ".Skybest.WMS.Auth",
        "in": "cookie"
      }
    }
  },
  "security": [
    {
      "CookieAuth": []
    }
  ]
}